+++ src/components/app/index.tsx
@@ -1,7 +1,9 @@
-import { Flex } from "@chakra-ui/react";
+import { Flex, Stack } from "@chakra-ui/react";
 import MainContainer from "../main-container";
 import NavBar from "../nav-bar";
 import BottomNavBar from "../nav-bar/bottom-nav-bar";
+import TopBar from "../top-bar";
+import ScrollContextProvider from "./scroll-context";
 
 function App() {
   return (
@@ -10,19 +12,24 @@ function App() {
       width="100vw"
       height="100dvh"
       paddingTop={"env(safe-area-inset-top)"}
-      // paddingBottom={"env(safe-area-inset-bottom)"}
       direction={{ base: "column", md: "row" }}
     >
-      <NavBar zIndex={10} hideBelow="md" minW="80px" height="100%" />
-      <MainContainer
-        height="100%"
-        maxW={"calc(100%-80px)"}
-        flex={1}
-        overflowX={"auto"}
-        padding={{ base: "0", md: "12px" }}
-        paddingLeft={{ base: "0", md: "80px" }}
-        marginLeft={{ base: "0", md: "-80px" }}
-      />
+      <ScrollContextProvider>
+        <NavBar zIndex={10} hideBelow="md" minW="80px" height="100%" />
+        <Stack
+          maxH="100%"
+          height="100%"
+          maxW={"calc(100%-80px)"}
+          flex={1}
+          overflowX={"auto"}
+          padding={{ base: "0", md: "12px" }}
+          paddingLeft={{ base: "0", md: "80px" }}
+          marginLeft={{ base: "0", md: "-80px" }}
+        >
+          <TopBar hideBelow="md" />
+          <MainContainer />
+        </Stack>
+      </ScrollContextProvider>
       <BottomNavBar hideFrom="md" />
     </Flex>
   );