+++ src/components/page/usePageScroll.ts
@@ -1,4 +1,0 @@
-import { useContext } from "react";
-import { PageContext } from "./page";
-
-export const usePageScroll = () => useContext(PageContext);