+++ src/components/table/paginated-table.tsx
@@ -6,7 +6,7 @@ import {
 } from "@/components/ui/pagination";
 import { For, HStack, Stack, Table } from "@chakra-ui/react";
 import { PropsWithChildren, useEffect, useMemo, useRef, useState } from "react";
-import { usePageScroll } from "../page/usePageScroll";
+import { useScroll } from "../app/useScroll";
 import PaginatedTableEmpty from "./paginated-table-empty";
 
 type Dict<T = any> = Record<string, T>;
@@ -20,7 +20,7 @@ function PaginatedTable<T extends string | number | Dict | undefined>({
 }>) {
   const scrollRef = useRef<HTMLDivElement>(null);
   const [page, setPage] = useState<number>(1);
-  const { onScroll } = usePageScroll();
+  const { onScroll } = useScroll();
 
   const pageSize = 40;
   const start = (page - 1) * pageSize;