+++ src/store/ui.ts
@@ -9,6 +9,7 @@ export enum ETabs {
   MyTracks = "/tracks",
   ShopGuide = "/checkout",
   About = "/about",
+  PrivacyPolicy = "/privacy-policy",
 }
 
 export const useUiStorePersist = create(